separate conftest.py for a little speed up
